+++ pkg/mimir/modules.go
@@ -189,9 +189,6 @@ func (t *Mimir) initRing() (serv services.Service, err error) {
 	if err != nil {
 		return nil, err
 	}
-
-	t.API.RegisterRing(t.Ring)
-
 	return t.Ring, nil
 }
 
@@ -447,7 +444,6 @@ func (t *Mimir) initIngesterService() (serv services.Service, err error) {
 
 func (t *Mimir) initIngester() (serv services.Service, err error) {
 	t.API.RegisterIngester(t.Ingester, t.Cfg.Distributor)
-
 	return nil, nil
 }